requirements.txt
streamlit
python-dotenv
groq
yt-dlp
youtube-transcript-api
PyPDF2
reportlab


